Change to qualifications of OpComm Director (#84)
* Change to qualifications of OpComm Director

The qualifications of the OpComm Director are now consistent with the rest of the Executive Board's.

* Fixed Issues with Chairman

Chairman is now a non-voting position, but still is the tie breaker. Secretary is now a non-voting member of eboard.

* Small Aesthetic Changes

i) Chairman, Secretary, Ad Hoc Director is now the order of non-Voting
Executive Board members.
ii) added (s) to Research and Development since there can be two
directors for the position.